NPC + DamageCanvas

+++ Assets/Scripts/Monster/Monster.cs
@@ -67,10 +67,10 @@ public class Monster : MonoBehaviour
     private int maxEnergy;
     private int maxHP;
     //stats increase Rate
-    private float ATK_INCREASE_RATE = 0.1f;
-    private float DEF_INCREASE_RATE = 0.1f;
-    private float HEAL_INCREASE_RATE = 0.05f;
-    private float ENERGY_INCREASE_RATE = 5;
+    private double ATK_INCREASE_RATE = 0.1;
+    private double DEF_INCREASE_RATE = 0.1;
+    private double HEAL_INCREASE_RATE = 0.05;
+    private double ENERGY_INCREASE_RATE = 5;
     public Monster(string ID,int atk,MonsterSO monsterSO){
         this.ATK = atk;
         this.ID = ID;
